refactor(view): lift the pane layout into a module of its own
Where the review's windows *are* was scattered through the file that decides what goes
in them. The panes and the predicates over them sat 1,200 lines above the before pane
that creates the second one, and the layout toggle sat below both -- so a question about
the split cost the file tree, delivery, navigation and the queue float to reach.

The seam is the one `6d542ed` cut for the queue float and #96 finished paying for: the
view is handed in, never required. A rebuilt pane needs the diff's keys and the cursor
entry point `view.cursor_moved` exists to be, and the toggle repaints -- all three arrive
as arguments, which is what keeps this module clear of the cycle `view` and `annotate`
sit in rather than routing it through a third file.

`V` stays one table and `view.lua` still owns it. This module takes `CRView` explicitly
and writes the before pane's handles onto it exactly where they were written before, the
way `syntax.lua` writes its caches there. No view state became a module local; the only
local here is the layout a reviewer chose, which was never on the view precisely because
it has to outlive the review it was chosen in.

Two names left this module's table and are re-exported so nothing outside learns of the
split: `keymaps.lua` binds `toggle_layout` to `gl` in both the diff and the tree, and
`annotate.lua` reads `focused_pane` to resolve a target. `goto_row` becomes a third
export in the other direction -- placing the panes is the layout's and following the diff
with the tree is the view's, so the one arrival both halves share stays glued where the
view that owns both is.

Nothing was renamed and nothing reworded. Every docstring travelled with the function it
documents, which is what lets the diff be read as a move.
