+++ lua/codereview/CLAUDE.md
@@ -34,13 +34,15 @@ change there is felt through.
 | `syntax.lua` | Treesitter harvest and replay onto the diff's rows, bounded by the viewport | stateful (extmarks) |
 | `types.lua` | Annotation types: defaults, normalisation, labels, and the directive that earns a type its keystroke | pure |
 | `view.lua` | The review view: buffers, windows, navigation, both layouts | stateful (windows) |
+| `view_layout.lua` | Where the review's windows are: the panes, the before pane, and the toggle between the unified and split layouts | stateful (windows) |
 
 ## How they stack
 
 - **Require nothing**: `diff`, `drafts`, `panel`, `queue`, `types`.
 - **Above them**, in that order: `git`, `payload`, `render`; then `state`, `config`; then
-  `archive`, `delivery`, `keymaps`, `syntax`; then `composer`, `hl`, `queue_float`.
-- **The hubs**: `view` requires thirteen of the modules above, `annotate` nine. A change that
+  `archive`, `delivery`, `keymaps`, `syntax`; then `composer`, `hl`, `queue_float`,
+  `view_layout`.
+- **The hubs**: `view` requires fourteen of the modules above, `annotate` nine. A change that
   is not local to a leaf almost certainly reaches one of them.
 - **On top**: `capture`, then `init`.
 
@@ -58,12 +60,15 @@ the `since-batch` scope. Function-local on both sides, as above. The alternative
 the snapshot in from outside, which would put the one scope that needs it into every caller
 of `resolve_scope` — and scope resolution is exactly what must stay one function.
 
-`keymaps` and `queue_float` would each be a third. Both run the view's exported actions,
-and both take them as an argument — `view` hands itself in — rather than requiring `view`
-for them. That is deliberate, and it is what leaves `keymaps` a function of its arguments
-and the configured annotation types. `queue_float` reads no view state either: the one
-field it needs, the window a float is open in, stays on `view` behind two accessors,
-because closing a float on submit is a rule about the view's windows.
+`keymaps`, `queue_float` and `view_layout` would each be a third. All three run the view's
+exported actions, and all three take them as an argument — `view` hands itself in — rather
+than requiring `view` for them. That is deliberate, and it is what leaves `keymaps` a
+function of its arguments and the configured annotation types. `queue_float` reads no view
+state either: the one field it needs, the window a float is open in, stays on `view` behind
+two accessors, because closing a float on submit is a rule about the view's windows.
+`view_layout` does read view state, but never its own copy of it: the `CRView` table is
+handed in beside the view and mutated in place, exactly as `syntax` and `state` are handed
+it, so there is still one table and `view` still owns it.
 
 ## Where reading pays
 
@@ -72,8 +77,9 @@ because closing a float on submit is a rule about the view's windows.
   `keymaps` is the one to open when the question is what a key does or where to add one —
   it is a table, not a search through the surface that happens to bind it.
 - **Carries the churn**: `view` and `annotate` by a wide margin, then `config`, `composer`
-  and `state`. `queue_float` is newly split out of `view` and inherits its history rather
-  than starting clean — the float's rows and its keys have both been rewritten recently.
+  and `state`. `queue_float` and `view_layout` are both newly split out of `view` and
+  inherit its history rather than starting clean — the float's rows and its keys have both
+  been rewritten recently, and the panes carry every trap the split layout ever cost.
 
 Re-derive with `git log --oneline --follow -- lua/codereview/<name>.lua` before leaning on
 that grouping; it moves.
